Added SPD converter
The 4.1 XSD does not seem to make use of all referenced elements
and is lacking in some respects to REDHAWK.  REDHAWK's implementation
block contains processor, os, runtime, and dependency, the last of
which helps point to softpkg and property dependencies.  This all
helps the framework collect assets at exectuable devices (and others)
that are friendly to the particular software component.  This may
cause an issue for any other imported framework elements.

Signed-off-by: Thomas Goodwin <btgoodwin@geontech.com>